# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PersonalRecordEventOut": {
        "properties": {
          "achieved_on": {
            "format": "date",
            "title": "Achieved On",
            "type": "string"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "display_precision": {
            "title": "Display Precision",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metric_display_name": {
            "title": "Metric Display Name",
            "type": "string"
          },
          "metric_key": {
            "title": "Metric Key",
            "type": "string"
          },
          "new_value": {
            "title": "New Value",
            "type": "number"
          },
          "player_display_name": {
            "title": "Player Display Name",
            "type": "string"
          },
          "player_id": {
            "format": "uuid",
            "title": "Player Id",
            "type": "string"
          },
          "previous_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Value"
          },
          "sample_size": {
            "title": "Sample Size",
            "type": "integer"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "player_id",
          "player_display_name",
          "metric_key",
          "metric_display_name",
          "unit",
          "display_precision",
          "previous_value",
          "new_value",
          "delta",
          "sample_size",
          "achieved_on",
          "source_status"
        ],
        "title": "PersonalRecordEventOut",
        "type": "object"
      },
...
      "SessionOut": {
        "properties": {
          "hit_count": {
            "default": 0,
            "title": "Hit Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pitch_count": {
            "default": 0,
            "title": "Pitch Count",
            "type": "integer"
          },
          "session_date": {
            "format": "date",
            "title": "Session Date",
            "type": "string"
          },
          "session_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Type"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "session_date",
          "source_status"
        ],
        "title": "SessionOut",
        "type": "object"
      },
      "SourceStatus": {
        "description": "Confidence level of vendor-supplied data.\n\nTrackMan can publish a session as PRELIMINARY and later republish it as\nVERIFIED with corrected values. Everything derived from an event carries the\nstatus of the data it was derived from so the UI can label it honestly.",
        "enum": [
          "PRELIMINARY",
          "VERIFIED"
        ],
        "title": "SourceStatus",
        "type": "string"
      },
...
      "TodaySnapshotOut": {
        "properties": {
          "athletes_trained": {
            "title": "Athletes Trained",
            "type": "integer"
          },
          "new_personal_records": {
            "title": "New Personal Records",
            "type": "integer"
          },
          "on_date": {
            "format": "date",
            "title": "On Date",
            "type": "string"
          },
          "sessions": {
            "title": "Sessions",
            "type": "integer"
          },
          "tracked_events": {
            "title": "Tracked Events",
            "type": "integer"
          }
        },
        "required": [
          "on_date",
          "athletes_trained",
          "sessions",
          "tracked_events",
          "new_personal_records"
        ],
        "title": "TodaySnapshotOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Internal player analytics for a baseball training facility. All athlete access is authorized server-side and scoped to one organization.",
    "title": "Baseball Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/dashboard/today": {
      "get": {
        "operationId": "today_api_v1_dashboard_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodaySnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Today",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/prs/recent": {
      "get": {
        "description": "Newest records across the organization -- the feed a coach scans first.",
        "operationId": "recent_records_api_v1_prs_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecordEventOut"
                  },
                  "title": "Response Recent Records Api V1 Prs Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Records",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Sessions Api V1 Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}